move Summary and SummaryTable to their component directories
